fix(rebase): Ensure robust interactive rebase conflict detection
Update the dev server endpoint to check both stdout and stderr for
`git rebase -i` conflict markers. This aligns with the Rust command's
behavior and prevents missed conflict detections due to `git` output
variations.

Additionally, this commit:
- Centralizes duplicated `branchSort` logic into a shared utility,
  establishing a consistent canonical ordering for all branch lists
  (e.g., app header, merge picker, rebase base picker).
- Adds explicit comments to `dev-server.mjs` and `ops.rs` to
  document the requirement of keeping `git branch --format` strings
  synchronized between the JS and Rust implementations.
